CHORE: rebase2 (#3424)
